3.1-drf-intro/smart_home/requirements.txt
asgiref==3.5.2
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
cryptography==37.0.4
distlib==0.3.4
Django==4.0.6
djangorestframework==3.13.1
fake-headers==1.0.2
filelock==3.6.0
greenlet==1.1.2
h11==0.13.0
html5lib==1.1
idna==3.3
iniconfig==1.1.1
mysqlclient==2.1.1
outcome==1.2.0
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.1
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytz==2022.2.1
requests==2.27.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.36
sqlparse==0.4.2
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.13.4
virtualenvwrapper-win==1.2.7
webencodings==0.5.1
wsproto==1.1.0
